Login Screen + Login View Model + Auth Repository +  Network Module + Auth Routes

+++ app/src/main/java/com/prototype/silver_tab/data/models/auth/AuthState.kt
@@ -5,7 +5,7 @@ data class AuthState(
     val isLoading: Boolean = false,
     val accessToken: String? = null,
     val refreshToken: String? = null,
-    val userId: Long? = null,
+    val userId: Int? = null,
     val username: String? = null,
     val email: String? = null,
     val role: Int? = null,
